# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca Virtual.",
        "description": "Essa API tem como objetivo o gerenciamento de livros em uma biblioteca virutal.",
        "contact": {
            "email": "peusodati@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3333/v1",
            "description": "API de Desenvolvimento"
        },
        {
            "url": "http://124.55.65.22:3333/v1",
            "description": "API de Produção"
        }
    ],
    "paths": {
...
        "/alunos/": {
            "post": {
                "sumary": "Cadastro de alunos",
                "description": "Esta rota é responsável por cadastrar um aluno",
                "tags": ["Alunos"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schema/Aluno"
                            },
                            "examples": {
                                "aluno": {
                                    "value": {
                                        "cursoId": "8a1dfb34-c487-404d-ae7b-8464da1679f0",
                                        "nome": "Pedro Henrique",
                                        "matricula": 1234
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schema/AlunoCurso"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação ou outro erro do cliente.",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "code": "tipo invalido",
                                            "expected": "string",
                                            "received": "undefined",
                                            "message": "required || O campo nome é obrigatório e deve conter pelo menos 4 caracteres.",
                                            "path": [
                                                "nome"
                                            ]
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe curso cadastro com esse Id. || Já existe aluno cadastrado com essa matricula"
                                        }
                                    } 
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "sumary": "Busca todos os alunos.",
                "description": "Esta rota é responsável por buscar todos os alunos cadastrados e o curso onde está matriculado.",
                "tags": ["Alunos"],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schema/AlunoCurso"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/alunos/{id}": {
            "get": {
                "sumary": "Busca de aluno por id.",
                "description": "Esta rota é responsável por buscar um aluno cadastrado pelo seu id.",
                "tags": ["Alunos"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id para buscar o aluno.",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "example": "8a1dfb34-c487-404d-ae7b-8464da1679f0"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schema/AlunoCurso"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "validation": "uuid",
                                            "code": "invalid_string",
                                            "message": "Invalid uuid || O campo id deve ser do tipo string uuid",
                                            "path": [
                                                "id"
                                            ]
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe aluno cadastro com esse Id."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "sumary": "Atualização de aluno por Id.",
                "description": "Esta rota é responsável por atualizar um aluno por Id.",
                "tags": ["Alunos"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id para atualizar o aluno.",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "example": "7a1dkc34-c487-404d-ae7b-8464da167888"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schema/Aluno"
                            },
                            "examples": {
                                "aluno": {
                                    "value": {
                                        "nome": "Joaquim",
                                        "matricula": 25265,
                                        "cursoId": "Opcional (uuid)"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aluno atualizado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schema/Aluno"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "validation": "uuid",
                                            "code": "invalid_string",
                                            "message": "Invalid uuid || O campo id deve ser do tipo string uuid",
                                            "path": [
                                                "id"
                                            ]
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe aluno cadastro com esse Id."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "sumary": "Exclusão de aluno por Id.",
                "description": "Esta rota é responsável por excluir um aluno por Id.",
                "tags": ["Alunos"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id para excluir o aluno.",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "example": "7a1dkc34-c487-404d-ae7b-8464da167888"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aluno excluído.",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "aluno": {
                                        "sumary": "Aluno excluído.",
                                        "value": {
                                            "message": "Aluno excluído com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "validation": "uuid",
                                            "code": "invalid_string",
                                            "message": "Invalid uuid || O campo id deve ser do tipo string uuid",
                                            "path": [
                                                "id"
                                            ]
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe aluno cadastro com esse Id."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schema": {
...
            "Aluno": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "matricula": {
                        "type": "integer"
                    },
                    "cursoId": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    }
                }
            },
            "AlunoCurso": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "matricula": {
                        "type": "integer"
                    },
                    "cursoId": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "Curso": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "nome": {
                                "type": "string"
                            },
                            "createdAt": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
        }
    }
}